xlsx import: parse short-hand version of range address
In order to indicate the entire row or column
(for the benefit of a spreadsheet program that may eventually
increase the supported number of rows or columns it supports)
a short-handed form of the range can be used
like "B:B" for the entire column,
or "1:10" for the first 10 rows of all columns.

Shorthand is nicer for humans
and allows for future expansion of row/column space

This reverts 24.8 commit a29d91ac403f1ed431ca95b8b9c290bd354c3ae7
    xlsx export: never export short-hand version of range address
which reduced the likelihood that we will ever see
such a range, because Excel seems to always
write out the full address.
However, the shorthand version is definitely valid for Excel.

Earlier patchsets checked that the new method and the old method
returned the same values for all existing unit tests.

Change-Id: I7677dfc207771a32222095416f728f7cd34a4f75
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/166570
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit cfb913db1b2024f8ff6a55f45742b303107a1924)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181766
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>