annot: add {add,remove}Annotation that don't notify
Add addAnnotationNoNotify method, that doesn't broadcast that an
annotation was added, and change addAnnotation to call the method
and in addition call the broadcast bits. Previously all this had
happened in the addAnnotation without the choice to not broadcast.

Change removeAnnotation in a similar way.

Change-Id: Ie15a386a8b8d4493d5b41fcbcb55924a693b46d2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168429
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
(cherry picked from commit e467cb52d68692e936eb07c2c1faa45f03dd1f82)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168573
Tested-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>